fix: CLI simulate and ExtNestDevice  (#207)
* fix: cli simulate - make exists flag not necessary if output_folder is not defined #204

* fix: Allow for list of values for Nest ExtNestDevice device #206

* fix: add MPI barrier for folder creation during BSB commands

* fix: add message explaining `exists` flag

* fix: use nest_parameter type handler for device constants.

* fix: add tests for cli bsb simulate command and test float list for NEST params

* fix: chunk size for NEST simulation test

* fix: split CLI unittests and use handle_command instead of subprocess

fix also MPI barrier

* fix: add MPI barrier before testing the counts of output nio files

+++ packages/bsb-core/bsb/cli/commands/_commands.py
@@ -15,6 +15,7 @@ from ...core import Scaffold, from_storage
 from ...exceptions import ConfigurationSyncError, NodeNotFoundError
 from ...option import BsbOption
 from ...reporting import report
+from ...services import MPI
 from ...storage import open_storage
 from . import BaseCommand
 
@@ -211,11 +212,21 @@ class BsbSimulate(BaseCommand, name="simulate"):
                 if name not in network.simulations and name == sim_name:
                     network.simulations[sim_name] = sim
         root = pathlib.Path(getattr(context.arguments, "output_folder", "./"))
-        try:
-            root.mkdir(exist_ok=context.arguments.exists)
-        except FileExistsError:
+        should_exit = False
+        if not MPI.get_rank():
+            try:
+                root.mkdir(
+                    exist_ok=context.arguments.exists
+                    or not hasattr(context.arguments, "output_folder")
+                )
+            except FileExistsError:
+                should_exit = True
+        should_exit = MPI.bcast(should_exit)
+        if should_exit:
             return report(
-                f"Could not create '{root.absolute()}', directory exists.", level=0
+                f"Could not create '{root.absolute()}', directory exists. "
+                "Use flag '--exists' to ignore this error.",
+                level=0,
             )
         try:
             result = network.run_simulation(sim_name)
@@ -250,29 +261,30 @@ class CacheCommand(BaseCommand, name="cache"):  # pragma: nocover
 
         from ...storage._util import _cache_path
 
-        if context.clear:
-            shutil.rmtree(_cache_path)
-            _cache_path.mkdir(parents=True, exist_ok=True)
-            print("Cache cleared")
-        else:
-            _cache_path.mkdir(parents=True, exist_ok=True)
-            files = [*_cache_path.iterdir()]
-            maxlen = 5
-            try:
-                maxlen = max(maxlen, max(len(file.name) for file in files))
-            except ValueError:
-                print("Cache is empty")
+        if not MPI.get_rank():
+            if context.clear:
+                shutil.rmtree(_cache_path)
+                _cache_path.mkdir(parents=True, exist_ok=True)
+                print("Cache cleared")
             else:
-                print(f"{'Files'.ljust(maxlen, ' ')}    Cached at\t\t\t    Size")
-                total_mb = 0
-                for f in files:
-                    name = f.name.ljust(maxlen, " ")
-                    stat = f.stat()
-                    stamp = datetime.fromtimestamp(stat.st_mtime)
-                    total_mb += (mb := stat.st_size / 1e6)
-                    line = f"{name}    {stamp}    {mb:.2f}MB"
-                    print(line)
-                print(f"Total: {total_mb:.2f}MB".rjust(len(line)))
+                _cache_path.mkdir(parents=True, exist_ok=True)
+                files = [*_cache_path.iterdir()]
+                maxlen = 5
+                try:
+                    maxlen = max(maxlen, max(len(file.name) for file in files))
+                except ValueError:
+                    print("Cache is empty")
+                else:
+                    print(f"{'Files'.ljust(maxlen, ' ')}    Cached at\t\t\t    Size")
+                    total_mb = 0
+                    for f in files:
+                        name = f.name.ljust(maxlen, " ")
+                        stat = f.stat()
+                        stamp = datetime.fromtimestamp(stat.st_mtime)
+                        total_mb += (mb := stat.st_size / 1e6)
+                        line = f"{name}    {stamp}    {mb:.2f}MB"
+                        print(line)
+                    print(f"Total: {total_mb:.2f}MB".rjust(len(line)))
 
     def get_options(self):
         return {
